% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{bivar_quali_quali}
\alias{bivar_quali_quali}
\title{bivar_quali_quali}
\usage{
bivar_quali_quali(variable_X, variable_Y, V_corrected = TRUE, na.rm = TRUE)
}
\arguments{
\item{variable_X}{Variable indépendante sous la forme "base$variable".}

\item{variable_Y}{Variable dépendante sous la forme "base$variable".}

\item{V_corrected}{A la valeur TRUE par défaut pour utiliser une version corrigée et donc plus précise du V de Cramer. À changer pour FALSE afin d'utiliser la version non corrigée.}

\item{na.rm}{A la valeur TRUE par défaut pour ne pas considérer les NA dans l'analyse. Changer cette valeur pour FALSE pourrait causer un message d'erreur. Si c'est le cas, rechanger pour TRUE.}
}
\value{
Retourne une liste contenant les mesures de significativité et de taille de l'effet suivantes, dans cet ordre : p-value et V de Cramer.
}
\description{
Fonction qui permet d'effectuer automatiquement l'analyse bivariée entre deux variables qualitatives.
}
