% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{bivar_quanti_quanti}
\alias{bivar_quanti_quanti}
\title{bivar_quanti_quanti}
\usage{
bivar_quanti_quanti(variable_x, variable_y, na.rm = TRUE)
}
\arguments{
\item{variable_x}{Variable indépendante sous la forme 'base$variable'.}

\item{variable_y}{Variable dépendante sous la forme 'base$variable'.}

\item{na.rm}{A la valeur 'TRUE' par défaut afin de retirer les NA du calcul. Peut être changée pour 'FALSE' afin de considérer les NA. Changer cette valeur pour 'FALSE' pourrait toutefois causer une erreur. Rechanger pour 'TRUE' si cela se produit.}
}
\value{
Retourne une liste contenant les mesures de significativité et de taille de leffet suivantes, dans cet ordre : p-value, r de Pearson (coefficient de corrélation) et R2 (coefficient de détermination).
}
\description{
Fonction qui permet d'effectuer automatiquement une analyse bivariée entre deux variables quantitatives continues. Si une variable quantitative est recodée en catégories, elle devient généralement une variable qualitative.
}
