% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{Importer}
\alias{Importer}
\title{Importer}
\usage{
Importer(indicateurs = NULL, colonnes = TRUE, encodage = "UTF-8")
}
\arguments{
\item{indicateurs}{Paramètre qui sert uniquement à l'importation de données provenant de la Banque Mondiale (package WDI). S'insère sous forme de vecteur de charactères.}

\item{colonnes}{A la valeur TRUE par défaut afin de nommer les colonnes de la base à partir de la première rangée de la source (changer pour colonnes = FALSE si la première rangée ne contient pas les noms des colonnes).}

\item{encodage}{A la valeur "UTF-8" par défaut afin de lire les caractères comme les accents. Il n'est pas recommandé d'en changer la valeur si vous ne comprenez pas ce qu'est l'encodage.}
}
\value{
Retourne un dataframe transformé en "tibble" pour en faciliter les manipulations.
}
\description{
Fonction interactive qui permet d'importer une base de données à partir d'un fichier, d'un URL ou de la fonction WDI() du package éponyme.
}
