% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{lm_eq}
\alias{lm_eq}
\title{lm_eq}
\usage{
lm_eq(x, y)
}
\arguments{
\item{x}{Variable indépendante (VI) sous la forme 'base$variable'}

\item{y}{Variable dépendante (VD) sous la forme 'base$variable'}
}
\value{
Équation de la droite en format 'character'.
}
\description{
Sert à imprimer une équation de droite de régression sur un graphique.
}
