% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{MargeErreur}
\alias{MargeErreur}
\title{MargeErreur}
\usage{
MargeErreur()
}
\value{
Retourne la marge d'erreur en nombre à virgule.
}
\description{
Fonction interactive qui permet de calculer la marge d'erreur d'un échantillon à partir du Z, du n et du p.
}
