% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{QualiQuanti}
\alias{QualiQuanti}
\title{QualiQuanti}
\usage{
QualiQuanti(variable_X, variable_Y, Y_ord = FALSE, na.rm = TRUE)
}
\arguments{
\item{variable_X}{Variable indépendante sous la forme "base$variable".}

\item{variable_Y}{Variable dépendante sous la forme "base$variable".}

\item{Y_ord}{A la valeur FALSE par défaut pour indiquer que la variable qualitative du croisement n'est pas ordinale (sans hiérarchie, nominale). Si la variable qualitative est ordinale (comme le niveau d'éducation), changer la valeur de l'Argument pour TRUE.}

\item{na.rm}{A la valeur TRUE par défaut pour ne pas considérer les NA dans l'analyse. Changer cette valeur pour FALSE pourrait causer un message d'erreur. Si c'est le cas, rechanger pour TRUE.}
}
\value{
Retourne une liste contenant nécessairement le p-value pour mesurer la significativité, puis la mesure de taille de l'effet correspondant au type de croisement : D de Cohen si VI dichotomique et tau de Kendall si VI polytomique ordinale.
}
\description{
Fonction qui permet d'effectuer automatiquement l'analyse bivariée entre une variable qualitative et une variable quantitative. Elle détecte automatiquement si la variable qualitative est dichotomique ou polytomique.
}
