% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{SommaireQuanti}
\alias{SommaireQuanti}
\title{SommaireQuanti}
\usage{
SommaireQuanti(variable, na.rm = TRUE)
}
\arguments{
\item{variable}{Variable quantitative continue sous la forme "base$variable"}

\item{na.rm}{A la valeur TRUE par défaut afin de retirer les NA du calcul. Peut être changée pour FALSE afin de considérer les NA.}
}
\value{
Retourne une liste contenant les mesures de tendance centrale et de dispersion suivantes, dans cet ordre : moyenne, médiane et écart-type.
}
\description{
Fonction qui permet d'effectuer l'analyse univariée d'une variable quantitative continue en une seule étape.
}
