% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{table_nomsvar}
\alias{table_nomsvar}
\title{table_nomsvar}
\usage{
table_nomsvar(variable_x, variable_y)
}
\arguments{
\item{variable_x}{Variable indépendante sous la forme 'base$variable'.}

\item{variable_y}{Variable dépendante sous la forme 'base$variable'.}
}
\value{
Retourne un tableau qui peut être sauvegardé dans un nouvel objet.
}
\description{
Fonction qui ajoute les noms des deux variables au 'table()'' afin de mieux identifier les colonnes et les rangées.
}
