% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{theme_pie}
\alias{theme_pie}
\title{theme_pie}
\usage{
theme_pie(couleur = "#018849", legende = TRUE)
}
\arguments{
\item{couleur}{Le vert de l'Université de Sherbrooke est la couleur par défaut, mais elle peut être changée manuelle pour toute autre couleur.}
}
\value{
Retourne le thème qui s'ajoute (+) à un graphique fait à partir ddu package ggplot2.
}
\description{
Thème personnalisé qui retire les fioritures dans la présentation d'un graphique en pointes de tarte, ou "pie chart".
}
