% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{univar_quali}
\alias{univar_quali}
\title{univar_quali}
\usage{
univar_quali(variable, na.rm = TRUE)
}
\arguments{
\item{variable}{Variable qualitative (ou catégorielle) sous la forme 'base$variable'}

\item{na.rm}{A la valeur 'TRUE' par défaut afin de retirer les NA du calcul. Peut être changée pour 'FALSE' afin de considérer les NA.}
}
\value{
Retourne le 'tableau des fréquences des valeurs' possibles que peut prendre la variable analysée.
}
\description{
Fonction qui effectue une analyse univariée sur une variable qualitative en une seule étape.
}
