% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{bivar_quali_quali}
\alias{bivar_quali_quali}
\title{bivar_quali_quali}
\usage{
bivar_quali_quali(variable_x, variable_y, v_corrige = TRUE, na.rm = TRUE)
}
\arguments{
\item{variable_x}{Variable indépendante sous la forme 'base$variable'.}

\item{variable_y}{Variable dépendante sous la forme 'base$variable'.}

\item{v_corrige}{A la valeur 'TRUE' par défaut pour utiliser une version corrigée et donc plus précise du V de Cramer. À changer pour 'FALSE' pour utiliser la version non corrigée.}

\item{na.rm}{A la valeur 'TRUE' par défaut afin de retirer les NA du calcul. Peut être changée pour 'FALSE' afin de considérer les NA.}
}
\value{
Retourne une liste contenant les mesures de significativité et de taille de leffet suivantes, dans cet ordre : p-value et V de Cramer.
}
\description{
Fonction qui effectue automatiquement une analyse bivariée entre deux variables qualitatives sélectionnées.
}
