% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{bivar_quanti_quanti}
\alias{bivar_quanti_quanti}
\title{bivar_quanti_quanti}
\usage{
bivar_quanti_quanti(variable_X, variable_Y, na.rm = TRUE)
}
\arguments{
\item{variable_X}{Variable indépendante sous la forme "base$variable".}

\item{variable_Y}{Variable dépendante sous la forme "base$variable".}

\item{na.rm}{A la valeur TRUE par défaut pour ne pas considérer les NA dans l'analyse. Changer cette valeur pour FALSE pourrait causer un message d'erreur. Si c'est le cas, rechanger pour TRUE.}
}
\value{
Retourne une liste contenant les mesures de significativité et de taille de l'effet suivantes, dans cet ordre : p-value, r de Pearson (coefficient de corrélation) et R2 (coefficient de détermination).
}
\description{
Fonction qui permet d'effectuer automatiquement l'analyse bivariée entre deux variables quantitatives continues. Si une variable quantitative est recodée en catégories, elle devient une variable qualitative.
}
