% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{marge_erreur}
\alias{marge_erreur}
\title{marge_erreur}
\usage{
marge_erreur(z = NULL, pr = NULL, n = NULL)
}
\value{
Retourne la valeur de la marge et
}
\description{
Fonction interactive qui permet de mesurer la marge derreur dun échantillon à partir du Z, du n et du p. Les valeurs des arguments peuvent être entrées non-interactivement aussi.
}
