% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{renommer}
\alias{renommer}
\title{renommer}
\usage{
renommer(base, anciens_noms, nouv_noms)
}
\arguments{
\item{base}{Objet qui représente la base de données.}

\item{anciens_noms}{Les noms des variables à renommer, sous la forme 'c("nom1", "nom2", "nom3")'.}

\item{nouv_noms}{Les nouveaux noms choisis pour les mêmes variables, dans le même ordre, toujours sous la forme 'c("nom1", "nom2", "nom3")'.}
}
\value{
Retourne le dataframe avec les modifications des noms des variables.
}
\description{
Fonction qui utilise 'rename()' mais qui simplifie son utilisation pour son utilisateur·ice.
}
