% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POL.R
\name{theme_pie}
\alias{theme_pie}
\title{theme_pie}
\usage{
theme_pie(couleur = "#018849", legende = TRUE)
}
\arguments{
\item{couleur}{Le vert 'Université de Sherbrooke' est la couleur par défaut, mais elle peut être changée manuellement pour toute autre couleur.}
}
\value{
Retourne le thème qui est ajouté (+) à un graphique fait à partir du package 'ggplot2'.
}
\description{
Thème personnalisé qui retire les fioritures dans un graphique en pointes de tarte, ou 'pie chart'.
}
